+++ source/Internal/Framework/Module/Facade/ActiveModulesDataProviderInterface.php
@@ -9,6 +9,8 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Facade;
 
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
+
 interface ActiveModulesDataProviderInterface
 {
     /**
@@ -20,4 +22,9 @@ interface ActiveModulesDataProviderInterface
      * @return string[]
      */
     public function getModulePaths(): array;
+
+    /**
+     * @return Template[][]
+     */
+    public function getTemplates(): array;
 }